When the language is set to Chinese,
the method of converting simplified Chinese to traditional Chinese is changed to using the method provided by zhconv.

When the "--merge_subtitle_with_sources" argument is enabled,
the video file extension will be appended with the subtitle language information, such as .en, .zh, .jp, etc.

The downloaded YouTube videos will be added with the format [vcodec^=avc1].